Use Cocoa APIs for font detection on OS X. Moving FontConfig to Linux platform only.
